requirements.txt
streamlit
langchain
openai
google-generativeai
PyMuPDF
python-dotenv
faiss-cpu
